# Use specific types when reading offset table