#### What this PR does

Modified the `index.ReadOffsetTable` function to use a specific type instead of a slice, that escapes to heap (for every label value in the postings offset table).

I will upstream this change to Prometheus, and then use the upstreamed version again when it's available, but wanted to bring it here first, especially because we already have benchmarks for this.

```
name          old time/op    new time/op    delta
BinaryReader    42.7µs ± 3%    33.9µs ± 8%  -20.63%  (p=0.008 n=5+5)

name          old alloc/op   new alloc/op   delta
BinaryReader    41.7kB ± 0%    35.3kB ± 0%  -15.31%  (p=0.000 n=5+4)

name          old allocs/op  new allocs/op  delta
BinaryReader       623 ± 0%       423 ± 0%  -32.10%  (p=0.008 n=5+5)
```

I also added a benchmark with a larger block size:

```
name                                old time/op    new time/op    delta
BinaryReader_LargerBlock/benchmark    2.72ms ± 4%    1.89ms ± 3%  -30.51%  (p=0.008 n=5+5)

name                                old alloc/op   new alloc/op   delta
BinaryReader_LargerBlock/benchmark    1.70MB ± 0%    1.06MB ± 0%     ~     (p=0.079 n=4+5)

name                                old allocs/op  new allocs/op  delta
BinaryReader_LargerBlock/benchmark     40.2k ± 0%     20.1k ± 0%  -49.93%  (p=0.008 n=5+5)
```

#### Which issue(s) this PR fixes or relates to

None, just was surfing the code.

#### Checklist

- [ ] Tests updated
- [ ] Documentation added
- [ ] `CHANGELOG.md` updated - the order of entries should be `[CHANGE]`, `[FEATURE]`, `[ENHANCEMENT]`, `[BUGFIX]`